macos: remove language standard c2x
Unfortunately, git 2.46.0 will fail to build on macOS Sequoia if the
language standard is set to c2x.

    CC reflog.o
reflog.c:213:24: error: too many arguments provided to function-like macro invocation
  213 | static int unreachable(struct expire_reflog_policy_cb *cb, struct commit *commit, struct object_id *oid)
      |                        ^
/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/lib/clang/16/include/__stddef_unreachable.h:18:9: note: macro 'unreachable' defined here
   18 | #define unreachable() __builtin_unreachable()
      |         ^
reflog.c:213:23: error: expected ';' after top level declarator

This happens, because the following header file defines "unreachable" as
a macro.

/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/lib/clang/16/include/__stddef_unreachable.h

It is unclear why the language standard is affecting this. It is also
unclear if other macOS versions beside Sequoia would be experiencing the
same issue. Furthermore, it is unknown why this was not an issue before
now.